+++ lib/repo_mgr/config.rb
@@ -15,7 +15,9 @@ module RepoMgr
       FileUtils.mkdir_p @cfg_dir
 
       @cfg_file = "#{@cfg_dir}/repo-mgr.yml"
-      File.write @cfg_file, { repos: {} }.to_yaml unless File.exist? @cfg_file
+      unless File.exist? @cfg_file
+        File.write @cfg_file, { repos: {}, packages: {} }.to_yaml
+      end
 
       @cfg = YAML.load_file @cfg_file
     end
@@ -24,16 +26,23 @@ module RepoMgr
       File.write @cfg_file, @cfg.to_yaml
     end
 
-    def upsert_repo(name, type, path, keyid)
+    def upsert_repo(options)
+      name = options[:name]
+      type = options[:type]
+
       if @cfg[:repos][name] && @cfg[:repos][name][:type] != type
         Tools.error "unable to change type for #{name} repository"
       end
 
       @cfg[:repos][name] = {
         type: type,
-        path: path,
-        keyid: keyid
+        path: options[:path],
+        keyid: options[:keyid]
       }
+
+      if options[:publisher]
+        @cfg[:repos][name][:publisher] = options[:publisher]
+      end
 
       save
     end
@@ -43,14 +52,15 @@ module RepoMgr
         Tools.error "unable to add packages to #{repo} - repo does not exist"
       end
 
-      @cfg[:repos][repo][:packages] ||= []
+      @cfg[:packages] ||= {}
+      @cfg[:packages][repo] ||= []
       pkg = File.basename path
 
-      if @cfg[:repos][repo][:packages].include?(pkg)
+      if @cfg[:packages][repo].include?(pkg)
         Tools.error "you already have #{pkg} in your #{repo} repo"
       end
 
-      @cfg[:repos][repo][:packages] << pkg
+      @cfg[:packages][repo] << pkg
 
       save
     end
@@ -61,9 +71,10 @@ module RepoMgr
           '- repo does not exist'
       end
 
-      @cfg[:repos][repo][:packages] ||= []
+      @cfg[:packages] ||= {}
+      @cfg[:packages][repo] ||= []
       pkg = File.basename path
-      @cfg[:repos][repo][:packages].delete pkg
+      @cfg[:packages][repo].delete pkg
 
       save
     end